pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["leeceo97 <leeceo97@naver.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
django = "^4.2.5"
djangorestframework = "^3.14.0"
isort = "^5.12.0"
black = "^23.9.1"
flake8 = "^6.1.0"
pre-commit = "^3.4.0"
django-environ = "^0.11.2"
drf-yasg = "^1.21.7"
pyjwt = "^2.8.0"
djangorestframework-camel-case = "^1.4.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
